+++ game/tile.py
@@ -1,4 +1,12 @@
 class Tile:
-    def __init__(self,letter, value):
+    def __init__(self, letter, value):
         self.letter = letter
         self.value = value
+
+    def __repr__(self):
+        return f"Tile(letter='{self.letter}', value={self.value})"
+
+    def __eq__(self, other):
+        if isinstance(other, Tile):
+            return self.letter == other.letter and self.value == other.value
+        return False