+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+0.0.3 - 2023-08-21
+ADDED:
+*Player class
+*Score for players
+*The tiles that each player has in possession
+*Players can swap pieces they don't want with a random one from the bag
+*Board class 
+*The board is a grid of squares
+*You can place tiles on the board
+FIXED:
+*Square class now has a default multiplier of x1
+*Tiles were given a custom comparison operator
+*Changed structure of the project, the dictionary is now in a separate folder
 0.0.2 - 2023-08-19
 ADDED:
 *List of words in Spanish considered legal for this version of Scrabble. 
@@ -5,6 +18,6 @@ ADDED:
 *Square class for future implementation of a Board class.
 0.0.1 - 2023-08-18
 ADDED:
-*Dictionary of letters to the Tilebag class. The amount of letters is now fixed, as well as their score.
+*List of letters to the Tilebag class. The amount of letters is now fixed, as well as their score.
 FIXED:
 *Fixed tests in order to stay relevant and evaluate the new format of the Tilebag class